Fix marshalling of `false` value in KeycloakBoolQuoted
Resolves #492

`false` is currently marshalled to `""` (empty string). This is incorrect: `""` represents the default value for an attribute which hasn't been explicitly set.
`""` could be interpreted as either `true` or `false` depending on what the default value is in Keycloak for the attribute considered.
As a result, `""` and `"false"` are not equivalent and `"false"` must be used explicitly to accurately turn off Keycloak attributes and not rely on default Keycloak behaviours.